utils: move argparse to dedicate file (they're too long)
